disp: msm: dp: add tiled dp-mst mode support
When 8K monitor is connected, dp-mst driver need to be configured
to tiled mode that two dp-mst stream will work together for super
frame display. A super encoder and brige are created to drive
two dp-mst sub-bridges, and dp-mst driver will select the super
encoder when 8K mode is selected.

Change-Id: I1ba8797675ff82048c854ac3294425dd558f90ec
Signed-off-by: Xiaowen Wu <wxiaowen@codeaurora.org>
Signed-off-by: Abhinav Kumar <abhinavk@codeaurora.org>