disp: msm: dp: set connected status for second stream
In the case of two DP MST connectors operating in bonding mode,
since we know that there will be two streams operating together
set the connected status when the second or the connector which is
detected later.

This helps the userspace to get the EDID for the combined stream
as the userspace would have otherwise ignored the second hotplug
event.

Change-Id: I24ed4983a100af68d4d3bb19e947006a7be29199
Signed-off-by: Abhinav Kumar <abhinavk@codeaurora.org>