disp: msm: sde: use new state when checking conn-crtc association
Currently during CRTC atomic check phase, driver is counting the
number of connectors attached to the CRTC by iterating overall
the connectors available and checking if they are attached to the
CRTC. The current implementation uses the old connector state,
which has an invalid state for the first commit.
This causes the number of attached connectors to always be 0 for
the first commit.

This change extracts the new connector state from the atomic state
and ensures the calculation is done before checking crtc features.

Change-Id: I58d641086f18e8624cbc2d432443323a6a44792e
Signed-off-by: Amine Najahi <anajahi@codeaurora.org>