disp: msm: dp: Extend mode filtering to support 8K
Currently DP driver determines if a mode is DSC capable based
on a DTSI entry and the required number of DSC to support it.
This approach does not scale when there is an overlap in DSC
requirement between DSI displays and external DP display,
thus causing one of the display to report modes that cannot
be supported.
This change compares the resources reserved for DP driver
calculated at initialization time and the currently available
ones to determine the correct number of resources that DP driver
can use. It also adds DSC and topology filtering logic and moves
DSC hardware specific from DP driver to SDE driver.

Change-Id: I8e601de33422b7c6d786826f7bfe152c4af8a6b5
Signed-off-by: Amine Najahi <anajahi@codeaurora.org>
Signed-off-by: Prashant Beniwal <pbeniw@codeaurora.org>