disp: msm: dp: expand dsc range_bpg_offset to support 10bpc
DSC driver is applying the range_bpg_offset values for 8 bpc/8bpp
to all compression ratios. This results in invalid DSC configuration
in 10bpc mode and sink is not able to decode this image resulting in
black screen.

This change adds a table of range_bpg_offset values to capture the
configuration for different bpc/bpp combinations.

Change-Id: I27e8edcbded8a8e512315599d768750ba473bd60
Signed-off-by: Amine Najahi <anajahi@codeaurora.org>
Signed-off-by: Prashant Beniwal <pbeniw@codeaurora.org>