disp: msm: dp: fix bpp capability check for compressed mode
Currently driver does not take into account the compression
ratio when determining the maximum supported bit per pixels for
a specific mode. This causes modes to publish a to a lower BPP
than the one it can support.

This change adds the DSC compression ratio to the data rate
calculation.

Change-Id: I8352a70947356f2996bb95e3f088064fa3da2a8f
Signed-off-by: Amine Najahi <anajahi@codeaurora.org>
Signed-off-by: Prashant Beniwal <pbeniw@codeaurora.org>